Proposed change to random initialisation of beam ray radial co-ordinates to eliminate bias
Current solution relies on the addition of 2 distributions, these are then corrected via a filter (2 - u[u > 1], where u is the array of results).
This leads to bias towards a value of unity, division by 2 instead of a simple filter just shifts bias to 0.5, suggested to just use 1 distribution with range 0 -> 1 instead, is there a reason 2 are needed?